fixed disparity between username/password in LFM and Client when updating user and pass

+++ Server.py
@@ -108,7 +108,8 @@ def change_username(username, password, new_usr):
     command = "UPDATE accounts SET usr = '%s' WHERE usr = '%s'" %(new_usr, username)
     c.execute(command)
 
-    os.rename(path + '/' + username, path + '/' + new_usr)
+    if os.path.exists(path + '/' + username):
+        os.rename(path + '/' + username, path + '/' + new_usr)
 
     return 'success'
